refactor: split code from Header component to utils and hooks
PT-1894.
